pyproject.toml
[tool.poetry]
name = "repl_python3_Flask"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
Flask = "^2.0.2"
bs4 = "^0.0.1"
pandas = "^1.4.1"
newsapi-python = "^0.2.6"
newsapi = "^0.1.1"
APScheduler = "^3.9.0"
yfinance = "^0.1.70"
waitress = "^2.0.0"
tqdm = "^4.62.3"
python-dotenv = "^0.19.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
